Tidied GA and Plausible recollection sources
