+++ package.json
@@ -23,7 +23,8 @@
     "react-dom": "^19.0.0",
     "react-ga4": "^2.1.0",
     "react-icons": "^5.5.0",
-    "react-type-animation": "^3.2.0"
+    "react-type-animation": "^3.2.0",
+    "script": "^0.0.12"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
