+++ pages/_app.js
@@ -1,7 +1,6 @@
 import { Analytics } from '@vercel/analytics/react';
 import { SpeedInsights } from '@vercel/speed-insights/next';
-// Add Google Analytics
-import Script from 'next/script';
+
 // Add custom event tracking
 import { useEffect } from 'react';
 import { useRouter } from 'next/router';
@@ -14,11 +13,6 @@ function MyApp({ Component, pageProps }) {
   // Track page views and user navigation
   useEffect(() => {
     const handleRouteChange = (url) => {
-      // Track page views (fixed the "nm" typo)
-      window.gtag('config', process.env.NEXT_PUBLIC_GA_ID, {
-        page_path: url,
-      });
-
       // Custom event for Vercel Analytics
       if (typeof window !== 'undefined' && window.va) {
         window.va('event', {
@@ -36,24 +30,6 @@ function MyApp({ Component, pageProps }) {
 
   return (
     <>
-      {/* Google Analytics Script */}
-      <Script
-        strategy="afterInteractive"
-        src={`https://www.googletagmanager.com/gtag/js?id=${process.env.NEXT_PUBLIC_GA_ID}`}
-      />
-      <Script
-        id="google-analytics"
-        strategy="afterInteractive"
-        dangerouslySetInnerHTML={{
-          __html: `
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-            gtag('config', process.env.NEXT_PUBLIC_GA_ID);
-          `,
-        }}
-      />
-
       <Component {...pageProps} />
       <Analytics
         mode="production"
